refactor: RequireAuth의 권한 제어 라우팅 로직을 middleware로 이관 → 인증 SSR 활성화
- RequireAuth 역할 변경: 권한 제어 → 쿠키 기반 유저 데이터 hydration
- RequireAuth의 변경된 역할에 따라 컴포넌트명 변경: 'AuthProvider'
